Success at parallelizsation, somewhat awkward but seems to work Big performance gains
